AI-model/python/requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
anyio==4.2.0
api==0.0.7
asgiref==3.7.2
astunparse==1.6.3
async==0.6.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bertopic==0.14.1
bidict==0.23.1
blinker==1.7.0
boto3==1.34.91
botocore==1.34.91
brotlipy==0.7.0
CacheControl==0.14.0
cached-property==1.5.2
cachetools==5.3.0
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cmd2==2.4.3
colorama==0.4.6
contourpy==1.0.7
cors==1.0.1
cryptography==39.0.0
cycler==0.11.0
Cython==0.29.34
decorator==4.4.2
exceptiongroup==1.2.0
ffmpeg==1.4
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-Sockets==0.2.1
flatbuffers==23.1.21
fonttools==4.39.4
frozenlist==1.4.1
fsspec==2023.4.0
future==1.0.0
gast==0.4.0
gevent==24.2.1
gevent-websocket==0.10.1
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.16.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.46.3
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
h5py==3.6.0
hdbscan==0.8.29
hpack==4.0.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.14.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.5
matplotlib==3.7.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
networkx==3.1
nltk==3.8.1
nose==1.3.7
numba==0.56.4
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pafy==0.5.5
pandas==1.5.3
pillow==10.3.0
pip==24.0
plotly==5.14.1
pluggy==1.0.0
priority==2.0.0
proglog==0.1.10
protobuf==4.25.2
py-hanspell==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.4
pycparser==2.21
pyfy==2.2.0
pynndescent==0.5.10
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
pytchat==0.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytube==15.0.0
pytz==2022.7.1
PyYAML==6.0
Quart==0.19.4
regex==2023.3.23
rel==0.4.9.8
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.10.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.98
setuptools==65.5.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
snscrape==0.6.2.20230321.dev3+g3dd9c28
soupsieve==2.4.1
sympy==1.11.1
taskgroup==0.0.0a4
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-macos==2.11.0
tensorflow-metal==0.7.0
termcolor==2.2.0
threadpoolctl==3.1.0
tldextract==5.1.2
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.66.2
transformers==4.28.1
typing_extensions==4.4.0
umap-learn==0.5.3
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
websocket==0.2.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
wheel==0.38.4
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.4
youtube-dl==2021.12.17
youtube-transcript-api==0.6.2
zope.event==5.0
zope.interface==6.3
zstandard==0.19.0
